Core: Move focusable into its own module
